requirements.txt
Flask
tensorflow
scikit-learn
numpy
matplotlib
pandas
python-dotenv gunicorn
rasterio
gdal